model_training/requirements.txt
datasets==2.2.2
numpy==1.22.2
pandas==1.4.1
PyYAML==6.0
scikit_learn==1.2.1
torch==1.11.0
transformers==4.17.0
